requirements.txt
annotated-types==0.7.0
anyio==4.6.0
Brotli
cachetools==5.5.0
certifi
charset-normalizer
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
distro==1.9.0
etelemetry==0.3.1
filelock
fitz==0.0.1.dev2
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
groq==0.11.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna
isodate==0.6.1
Jinja2
looseversion==1.3.0
lxml==5.3.0
MarkupSafe
mkl-service==2.4.0
mkl_fft
mkl_random
mpmath
networkx
nibabel==5.3.0
nipype==1.8.6
numpy
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
pillow
proto-plus==1.24.0
protobuf==5.28.2
prov==2.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
PyMuPDF==1.24.11
pyparsing==3.1.4
PySocks
python-dateutil==2.9.0.post0
pytz==2024.2
pyxnat==1.6.2
PyYAML
rdflib==6.3.2
requests
rsa==4.9
scipy==1.14.1
setuptools==72.1.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sympy
termcolor==2.5.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
traits==6.3.2
typing_extensions
tzdata==2024.2
uritemplate==4.1.1
urllib3
wheel==0.44.0
win-inet-pton
